added student IDs to USERAGENT
Set the USERAGENT variable in Config.ini so that it contains all students' IDs separated by a comma (the numbers! e.g. IR UF24 123123213,12312312,123123  ) of the group members. If you fail to do this properly, your crawler will not exist in the server's log, which will put your grade for this project at risk.